SLING-12734: Move AliasHandler into new class
